Add entites and migrations files for services. Add publish ack.
